remove kobo api
